Revisao CR

+++ models/productsModel.js
@@ -1,21 +1,39 @@
 const { ObjectId } = require('mongodb');
 const connection = require('./connections');
 
-const getAll = async () => (
-  connection()
-    .then((db) => db.collection('products').find().toArray())
-    .then((products) =>
-      products.map(({ _id, name, quantity }) => ({
-        id: _id,
-        name,
-        quantity,
-      })))
-);
+const getAll = async () => {
+  try {
+    const allProducts = await connection()
+      .then((db) => db.collection('products').find().toArray())
+      .then((products) =>
+        products.map(({ _id, name, quantity }) => ({
+          id: _id,
+          name,
+          quantity,
+        })));
+    return allProducts;
+  } catch (err) {
+    console.error(err);
+    return false;
+  }
+};
 
 const getProductById = async (id) => {
+  if (!ObjectId.isValid(id)) return 'not_found';
   try {
     const product = await connection()
       .then((db) => db.collection('products').findOne({ _id: ObjectId(id) }));
+    return product || [];
+  } catch (err) {
+    console.error(err);
+    return false;
+  }
+};
+
+const getProductByName = async (name) => {
+  try {
+    const product = await connection()
+      .then((db) => db.collection('products').findOne({ name }));
     return product || [];
   } catch (err) {
     console.error(err);
@@ -35,6 +53,7 @@ const createProducts = async (product) => {
 };
 
 const deleteProduct = async (id) => {
+  if (!ObjectId.isValid(id)) return 'not_found';
   try {
     const { result } = await connection()
       .then((db) => db.collection('products').deleteOne({ _id: ObjectId(id) }));
@@ -46,6 +65,7 @@ const deleteProduct = async (id) => {
 };
 
 const updateProduct = async (id, updatedProduct) => {
+  if (!ObjectId.isValid(id)) return 'not_found';
   try {
     const { result } = await connection()
       .then((db) => db.collection('products').updateOne({ _id: ObjectId(id) }, { $set: updatedProduct }));
@@ -57,9 +77,12 @@ const updateProduct = async (id, updatedProduct) => {
 };
 
 const repeatedIds = async (arrayIds) => {
+  const isValidIds = arrayIds.every((id) => ObjectId.isValid(id));
+  if (!isValidIds) return 'not_found';
+  const validIds = arrayIds.map((id) => ObjectId(id));
   try {
     const arrayOfMatchIds = await connection()
-      .then((db) => db.collection('products').find({ _id: { $in: arrayIds } }).toArray());
+      .then((db) => db.collection('products').find({ _id: { $in: validIds } }).toArray());
     return arrayOfMatchIds;
   } catch (err) {
     console.error(err);
@@ -74,4 +97,5 @@ module.exports = {
   deleteProduct,
   updateProduct,
   repeatedIds,
+  getProductByName,
 };
